+++ src/main/kotlin/com/github/sleepypanda/feesh/settings/categories/Alerts.kt
@@ -254,7 +254,7 @@ object Alerts : CategoryKt("Alerts") {
 
     var alertOnRainEndingSoon by boolean(false) {
         this.name = Translated("Alert when Rain/Thunder/Blizzard ends soon")
-        this.description = Translated("${GRAY}Shows a title and plays a sound when active Rain/Thunder/Blizzard ends soon. It's applicable to The Park, Spider's Den, Lotus Atoll, and Jerry's Workshop. Please enable ${YELLOW}TabList settings -> General Info widget -> Show Rain / Show Blizzard")
+        this.description = Translated("${GRAY}Shows a title and plays a sound when active Rain/Thunder/Blizzard ends soon. It's applicable to The Park, Spider's Den, Lotus Atoll, Backwater Bayou, and Jerry's Workshop. Please enable ${YELLOW}TabList settings -> General Info widget -> Show Rain / Show Blizzard")
     }
 
     init {